+++ src/runtime/Projects.tsx
@@ -1,3 +1,5 @@
-export function Projects() {
-  return <svg aria-hidden="true" viewBox="0 0 300 150" preserveAspectRatio="xMidYMid meet"><path fill="currentColor" className="blinking" d="M180 122h-33c-3 0-5-2-5-5s2-6 5-6h32c3 0 5 3 5 6s-2 5-4 5Z" /><path fill="currentColor" d="M97 126c-2 0-4-1-5-2-2-3-1-7 2-8l57-41-35-37c-2-2-2-5 0-7s5-2 7 0l38 41c1 1 2 2 2 4v1c0 2-1 4-2 5l-61 44c-1 0-2 0-3 0Z" /></svg>
+import type { SVGProps } from 'react'
+
+export function Projects(props: SVGProps<SVGSVGElement>) {
+  return <svg {...props} viewBox="0 0 300 150" preserveAspectRatio="xMidYMid meet"><path fill="currentColor" className="blinking" d="M180 122h-33c-3 0-5-2-5-5s2-6 5-6h32c3 0 5 3 5 6s-2 5-4 5Z" /><path fill="currentColor" d="M97 126c-2 0-4-1-5-2-2-3-1-7 2-8l57-41-35-37c-2-2-2-5 0-7s5-2 7 0l38 41c1 1 2 2 2 4v1c0 2-1 4-2 5l-61 44c-1 0-2 0-3 0Z" /></svg>
 }